mv remind specific out of generic
